Update browser tests and add badges

+++ webdriver.bash
@@ -1,11 +1,9 @@
 #!/bin/bash
 
 set -euo pipefail
-pkill -f selenium-standalone || true
 pid=""
 
 cleanup() {
-	pkill -f selenium-standalone || true
 	if [ "$pid" != "" ]
 	then
 		kill "$pid" || true
@@ -14,9 +12,23 @@ cleanup() {
 trap "cleanup" EXIT INT
 PATH="$PATH:./node_modules/.bin/"
 
-selenium-standalone install --silent
-selenium-standalone start -- -log /tmp/protractor.log &
-sleep 2
-pid="$!"
-echo "node webdriver"
-node webdriver.js
+if [ "$BROWSER" != "SAUCELABS" ]
+then
+	if netstat -tnlp | grep --color -E 4444 >/dev/null
+	then
+		echo "Using existing selenium"
+	else
+		selenium-standalone install --silent
+		selenium-standalone start -- -log /tmp/protractor.log &
+		pid="$!"
+	fi
+	node webdriver.js
+	exit "$?"
+fi
+
+set +e
+
+browserName="chrome" platform="Windows 10" version="58" node webdriver.js
+browserName="firefox" platform="Windows 10" version="55" node webdriver.js
+browserName="internet explorer" platform="Windows 10" version="11" node webdriver.js
+browserName="iphone" platform="Mac 10.11" version="10.2" node webdriver.js
